chore(test): move jit-html tests to runtime-html folder
