+++ packages/__tests__/3-runtime-html/binding-command.class.spec.ts
