+++ packages/__tests__/3-runtime-html/binding-resources.spec.ts
