+++ packages/__tests__/3-runtime-html/local-dependency-inheritance.spec.ts
